Update README.md
I left some comments in there that I set off with HTML tags <!--EM.

My one big problem is that I don't see where the reader is accessing COVID-19 information. Covid-19 isn't even mentioned until the second-to-last step where the reader names their chatbot. Is that part of the JSON file? We need to be mentioning that more if that's the heading in the tutorial. It's not clear to me reading these steps what this has to do with COVID-19. 

Also, note my question about Postman. That's just sort of randomly introduced. Should it be in the prereqs?